# Conflicts:
#	src/Main.cc
#	src/web/include/web/web.h
#	src/web/src/main.js
#	src/web/src/request_handler.cpp
#	src/web/src/request_handler.h
#	src/web/src/web_serve.cpp

+++ src/odb/src/db/dbTechLayerMinCutRule.h
@@ -49,7 +49,7 @@ class _dbTechLayerMinCutRule : public _dbObject
   int within_cut_dist;
   int length_;
   int length_within_dist_;
-  int area_;
+  int64_t area_;
   int area_within_dist_;
 };
 dbIStream& operator>>(dbIStream& stream, _dbTechLayerMinCutRule& obj);
